+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `markdown.marp.newMarpMarkdown` command to create empty Marp Markdown ([#255](https://github.com/marp-team/marp-vscode/pull/255))
 - Contribution to "New File…" in File menu and welcome screen _(Experimental: Required opt-in by `workbench.welcome.experimental.startEntries` in VS Code 1.58+)_ ([#255](https://github.com/marp-team/marp-vscode/pull/255))
+- Allow to disable math typesetting in Marp Markdown by `"markdown.marp.mathTypesetting": "off"` setting ([#256](https://github.com/marp-team/marp-vscode/issues/256), [#258](https://github.com/marp-team/marp-vscode/pull/258))
 
 ### Fixed